+++ internal/server/apparmor/archive.go
@@ -6,7 +6,6 @@ import (
 	"os/exec"
 	"path/filepath"
 	"strings"
-	"text/template"
 
 	"github.com/google/uuid"
 
@@ -14,33 +13,6 @@ import (
 	internalUtil "github.com/lxc/incus/v6/internal/util"
 	"github.com/lxc/incus/v6/shared/revert"
 )
-
-var archiveProfileTpl = template.Must(template.New("archiveProfile").Parse(`#include <tunables/global>
-profile "{{.name}}" {
-  #include <abstractions/base>
-  #include <abstractions/nameservice>
-
-{{range $index, $element := .allowedCommandPaths}}
-  {{$element}} mixr,
-{{- end }}
-
-  {{ .outputPath }}/ rw,
-  {{ .outputPath }}/** rwl,
-  {{ .backupsPath }}/** rw,
-  {{ .imagesPath }}/** r,
-
-  signal (receive) set=("term"),
-
-  # Capabilities
-  capability chown,
-  capability dac_override,
-  capability dac_read_search,
-  capability fowner,
-  capability fsetid,
-  capability mknod,
-  capability setfcap,
-}
-`))
 
 // ArchiveWrapper is used as a RunWrapper in the rsync package.
 func ArchiveWrapper(sysOS *sys.OS, cmd *exec.Cmd, output string, allowedCmds []string) (func(), error) {
