+++ internal/server/apparmor/qemuimg.go
@@ -12,45 +12,11 @@ import (
 	"path/filepath"
 	"strconv"
 	"strings"
-	"text/template"
 
 	"github.com/lxc/incus/v6/internal/server/sys"
 	"github.com/lxc/incus/v6/shared/ioprogress"
 	"github.com/lxc/incus/v6/shared/subprocess"
 )
-
-var qemuImgProfileTpl = template.Must(template.New("qemuImgProfile").Parse(`#include <tunables/global>
-profile "{{ .name }}" flags=(attach_disconnected,mediate_deleted) {
-  #include <abstractions/base>
-
-  capability dac_override,
-  capability dac_read_search,
-  capability ipc_lock,
-
-  /proc/sys/vm/max_map_count r,
-  /sys/devices/**/block/*/queue/max_segments  r,
-  /sys/devices/**/block/*/zoned  r,
-  /sys/devices/system/node r,
-  /sys/devices/system/node/** r,
-
-{{range $index, $element := .allowedCmdPaths}}
-  {{$element}} mixr,
-{{- end }}
-
-  {{ .pathToImg }} rk,
-
-{{- if .dstPath }}
-  {{ .dstPath }} rwk,
-{{- end }}
-
-{{if .libraryPath -}}
-  # Entries from LD_LIBRARY_PATH
-{{range $index, $element := .libraryPath}}
-  {{$element}}/** mr,
-{{- end }}
-{{- end }}
-}
-`))
 
 type nullWriteCloser struct {
 	io.Writer
